+++ csatar-plugins/csatar/models/PermissionBasedAccess.php
@@ -42,7 +42,7 @@ class PermissionBasedAccess extends Model
         return '\\' . static::class;
     }
 
-    public function getRightsForMandateTypes(array $mandateTypeIds = [], bool $own = false, bool $is2fa = false)
+    public function getRightsForMandateTypes(array $mandateTypeIds = [], bool $own = false, bool $is2fa = false, bool $ignoreCache = false)
     {
 
         $associationId = $this->getAssociationId();
@@ -59,17 +59,22 @@ class PermissionBasedAccess extends Model
             return;
         }
 
-        $rights = Db::table('csatar_csatar_mandates_permissions')
-            ->when(!$own, function ($query) {
-                return $query->where(
-                    function ($query) {
-                        return $query->where('own', '<>', 1)->orWhereNull('own');
-                    });
-            })
-            ->whereIn('mandate_type_id', $mandateTypeIds)
-            ->where('model', self::getModelName())
-            ->get();
+        if (!$ignoreCache) {
+            $rights = $this->getRightsForMandateTypesFromSession($mandateTypeIds, self::getModelName());
+        }
 
+        if ($ignoreCache || empty($rights) || $rights->count() == 0) {
+            $rights = MandatePermission::whereIn('mandate_type_id', $mandateTypeIds)
+               ->where('model', self::getModelName())
+               ->get();
+            $this->saveRightsForMandateTypesToSession($rights);
+        }
+
+        $rights = $rights->when(!$own, function ($collection) {
+            return $collection->reject(function ($item) {
+                return $item->own == 1;
+            });
+        });
         $rights = $rights->groupBy('field');
 
         return $rights->map(function ($item, $key) use ($is2fa) {
@@ -85,6 +90,15 @@ class PermissionBasedAccess extends Model
         });
     }
 
+    public function getRightsForMandateTypesFromSession(array $mandateTypeIds, string $model){
+        $sessionRecord = Session::get('scout.rightsForMandateTypes');
+        return $sessionRecord ? $sessionRecord->whereIn('mandate_type_id', $mandateTypeIds)->where('model', $model) : null;
+    }
+
+    public function saveRightsForMandateTypesToSession($rights){
+        Session::put('scout.rightsForMandateTypes', $rights);
+    }
+
     public function getGuestRightsForModel()
     {
         $associationId = $this->getAssociationId();
@@ -94,7 +108,10 @@ class PermissionBasedAccess extends Model
         $sessionRecord = Session::get('guest.rightsForModels');
         $sessionRecordForModel = $sessionRecord ? $sessionRecord->get($key) : null;
 
-        if (!empty($sessionRecordForModel) && $sessionRecordForModel['savedToSession'] >= RightsMatrix::getRightsMatrixLastUpdateTime() && $sessionRecordForModel['rights']->count() != 0) {
+        if (!empty($sessionRecordForModel)
+            && $sessionRecordForModel['savedToSession'] >= RightsMatrix::getRightsMatrixLastUpdateTime()
+            && $sessionRecordForModel['rights']->count() != 0)
+        {
             return $sessionRecordForModel['rights'];
         }
 
@@ -194,5 +211,25 @@ class PermissionBasedAccess extends Model
     public static function getOrganizationTypeModelNameUserFriendly()
     {
         return '';
+    }
+
+    public function getAssociation() {
+        return null;
+    }
+
+    public function getDistrict() {
+        return null;
+    }
+
+    public function getTeam() {
+        return null;
+    }
+
+    public function getTroop() {
+        return null;
+    }
+
+    public function getPatrol() {
+        return null;
     }
 }