+++ csatar-plugins/csatar/models/Scout.php
@@ -2,8 +2,11 @@
 
 use Auth;
 use Csatar\Csatar\Models\MandateType;
+use DateTime;
 use Db;
+use Flash;
 use Lang;
+use Log;
 use Session;
 use Model;
 use Csatar\Csatar\Classes\RightsMatrix;
@@ -195,30 +198,36 @@ class Scout extends OrganizationBase
      */
     public function filterFields($fields, $context = null) {
         // populate the Troop and Patrol dropdowns with troops and patrols that belong to the selected team
-        $fields->troop->options = [];
-        $team_id = $this->team_id;
-        if ($team_id) {
-            foreach (\Csatar\Csatar\Models\Troop::teamId($team_id)->get() as $troop) {
-                $fields->troop->options += [$troop['id'] => $troop['extendedName']];
+        if (isset($fields->troop)) {
+            $fields->troop->options = [];
+            $team_id = $this->team_id;
+            if ($team_id) {
+                foreach (\Csatar\Csatar\Models\Troop::teamId($team_id)->get() as $troop) {
+                    $fields->troop->options += [$troop['id'] => $troop['extendedName']];
+                }
             }
         }
 
         // populate the Patrol dropdown with patrols that belong to the selected team and to the selected troop
-        $fields->patrol->options = [];
-        $troop_id = $this->troop_id;
-        if ($troop_id) {
-            foreach (\Csatar\Csatar\Models\Patrol::troopId($troop_id)->get() as $patrol) {
-                $fields->patrol->options += [$patrol['id'] => $patrol['extendedName']];
+        if (isset($fields->patrol)) {
+            $fields->patrol->options = [];
+            $troop_id = $this->troop_id;
+            if ($troop_id) {
+                foreach (\Csatar\Csatar\Models\Patrol::troopId($troop_id)->get() as $patrol) {
+                    $fields->patrol->options += [$patrol['id'] => $patrol['extendedName']];
+                }
             }
-        }
-        else if ($team_id) {
-            foreach (\Csatar\Csatar\Models\Patrol::teamId($team_id)->get() as $patrol) {
-                $fields->patrol->options += [$patrol['id'] => $patrol['extendedName']];
+            else if ($team_id) {
+                foreach (\Csatar\Csatar\Models\Patrol::teamId($team_id)->get() as $patrol) {
+                    $fields->patrol->options += [$patrol['id'] => $patrol['extendedName']];
+                }
             }
         }
 
         // populate the Legal Relationships dropdown with legal relationships that belong to the selected team's association
-        $fields->legal_relationship->options = $this->team ? \Csatar\Csatar\Models\LegalRelationship::associationId($this->team->district->association->id)->lists('name', 'id') : [];
+        if (isset($fields->legal_relationship)) {
+            $fields->legal_relationship->options = $this->team ? \Csatar\Csatar\Models\LegalRelationship::associationId($this->team->district->association->id)->lists('name', 'id') : [];
+        }
     }
 
     public $fillable = [
@@ -400,6 +409,18 @@ class Scout extends OrganizationBase
         $this->patrol_id = $this->patrol_id != 0 ? $this->patrol_id : null;
     }
 
+    function beforeDelete()
+    {
+        $now = new DateTime();
+        $mandates = Mandate::where('scout_id', $this->id)->get();
+        foreach ($mandates as $mandate) {
+            if ($mandate->start_date < $now && ($mandate->end_date > $now || $mandate->end_date == null)) {
+                Flash::error(str_replace('%name', $this->getFullName(), Lang::get('csatar.csatar::lang.plugin.admin.scout.activeMandateDeleteError')));
+                return false;
+            }
+        }
+    }
+
     private function generateEcsetCode()
     {
         $team = Team::find($this->team_id);
@@ -520,6 +541,15 @@ class Scout extends OrganizationBase
         return $query->where('team_id', $id);
     }
 
+    public function scopeAssociations($query, array $associationIds)
+    {
+        return $query->whereHas('team', function ($query) use ($associationIds) {
+            $query->whereHas('district', function ($query) use ($associationIds) {
+                $query->whereIn('association_id', $associationIds);
+            });
+        });
+    }
+
     /*
      * Returns all the mandates scout has in a specific association
      */
@@ -589,8 +619,52 @@ class Scout extends OrganizationBase
         return $scoutMandateTypeIds;
     }
 
+    public function getMandateTypeIdsInOrganizationTree(PermissionBasedAccess $model): ?array
+    {
+        $modelAssociation = $model->getAssociation();
+        $modelDistrict = $model->getDistrict();
+        $modelTeam = $model->getTeam();
+        $modelTroop = $model->getTroop();
+        $modelPatrol = $model->getPatrol();
+
+        if (!empty($modelAssociation)) {
+            $mandateIdsForAssociation = $this->getMandatesForOrganization($modelAssociation)
+                                             ->pluck('mandate_type_id')->toArray();
+        }
+
+        if (!empty($modelDistrict)) {
+            $mandateIdsForDistrict = $this->getMandatesForOrganization($modelDistrict)
+                                          ->pluck('mandate_type_id')->toArray();
+        }
+
+        if (!empty($modelTeam)) {
+            $mandateIdsForTeam = $this->getMandatesForOrganization($modelTeam)
+                                      ->pluck('mandate_type_id')->toArray();
+        }
+
+        if (!empty($modelTroop)) {
+            $mandateIdsForTroop = $this->getMandatesForOrganization($modelTroop)
+                                       ->pluck('mandate_type_id')->toArray();
+        }
+
+        if (!empty($modelPatrol)) {
+            $mandateIdsForPatrol = $this->getMandatesForOrganization($modelPatrol)
+                                        ->pluck('mandate_type_id')->toArray();
+        }
+
+        return array_merge(
+            $mandateIdsForAssociation ?? [],
+            $mandateIdsForDistrict ?? [],
+            $mandateIdsForTeam ?? [],
+            $mandateIdsForTroop ?? [],
+            $mandateIdsForPatrol ?? [],
+            MandateType::getScoutMandateTypeIdInAssociation($modelAssociation->id)
+        );
+    }
+
     public function getMandatesForOrganization(PermissionBasedAccess $organization) {
         return $this->mandates()
+            ->where('mandate_model_type', $organization->getModelName())
             ->where('mandate_model_id', $organization->id)
             ->where('start_date', '<=', date('Y-m-d H:i'))
             ->where(function ($query) {
@@ -628,104 +702,9 @@ class Scout extends OrganizationBase
             $is2fa = true;
         }
 
-        $associationId  = $model->getAssociationId();
-        if ($model->getModelName() === '\Csatar\Csatar\Models\Scout' && !$isOwn) {
-            $modelAssociation = $model->getAssociation();
-            $modelDistrict = $model->getDistrict();
-            $modelTeam = $model->getTeam();
-            $modelTroop = $model->getTroop();
-            $modelPatrol = $model->getPatrol();
+        $mandateTypeIds = $this->getMandateTypeIdsInOrganizationTree($model);
 
-            if (!empty($modelAssociation)) {
-                $mandateIdsForAssociation = $this->getMandatesForOrganization($modelAssociation)
-                                          ->pluck('mandate_type_id')->toArray();
-            }
-
-            if (!empty($modelDistrict)) {
-                $mandateIdsForDistrict = $this->getMandatesForOrganization($modelDistrict)
-                                          ->pluck('mandate_type_id')->toArray();
-            }
-
-            if (!empty($modelTeam)) {
-                $mandateIdsForTeam = $this->getMandatesForOrganization($modelTeam)
-                                          ->pluck('mandate_type_id')->toArray();
-            }
-
-            if (!empty($modelTroop)) {
-                $mandateIdsForTroop = $this->getMandatesForOrganization($modelTroop)
-                                          ->pluck('mandate_type_id')->toArray();
-            }
-
-            if (!empty($modelPatrol)) {
-                $mandateIdsForPatrol = $this->getMandatesForOrganization($modelPatrol)
-                                          ->pluck('mandate_type_id')->toArray();
-            }
-
-            $mandateTypeIds = array_merge(
-                $mandateIdsForAssociation ?? [],
-                $mandateIdsForDistrict ?? [],
-                $mandateIdsForTeam ?? [],
-                $mandateIdsForTroop ?? [],
-                $mandateIdsForPatrol ?? [],
-                MandateType::getScoutMandateTypeIdInAssociation($associationId)
-            );
-            $ignoreCache = true;
-
-        } else {
-            $mandateTypeIds = $this->getMandateTypeIdsInAssociation($associationId, $this->updated_at, $ignoreCache);
-        }
-
-        $rightsForModel = $ignoreCache ? null : $this->getRightsForModelFromSession($model, $associationId, $isOwn);
-
-        if(empty($rightsForModel) || $rightsForModel->count() == 0) {
-            $rightsForModel = $model->getRightsForMandateTypes($mandateTypeIds, $isOwn, $is2fa);
-            $this->saveRightsForModelToSession($model, $rightsForModel, $associationId, $isOwn);
-        }
-
-        return $rightsForModel;
-    }
-
-    public function getRightsForModelFromSession($model, $associationId, $own = false) {
-        $sessionRecord = Session::get('scout.rightsForModels');
-
-        if(empty($sessionRecord) || empty($model) || empty($associationId)) {
-            return;
-        }
-
-        $key = $associationId . $model::getModelName() . ($own ? '_own' : '');
-
-        $sessionRecordForModel = $sessionRecord->get($key);
-
-        if (!empty($sessionRecordForModel)
-            && $sessionRecordForModel['savedToSession'] >= RightsMatrix::getRightsMatrixLastUpdateTime()
-            && isset($sessionRecordForModel['rights'])
-            && $sessionRecordForModel['rights']->count() != 0) {
-           return $sessionRecordForModel['rights'];
-        }
-
-        return null;
-    }
-
-    public function saveRightsForModelToSession($model, $rightsForModel, $associationId, $own) {
-        $modelName = $model::getModelName();
-        $key = $associationId . $modelName . ($own ? '_own' : '');
-        $sessionRecord = Session::get('scout.rightsForModels');
-
-        if(empty($sessionRecord)){
-            $sessionRecord = new Collection([]);
-        }
-
-        $sessionRecord = $sessionRecord->replace([
-            $key => [
-                'associationId' => $associationId,
-                'model' => $modelName,
-                'own' => $own,
-                'savedToSession' => date('Y-m-d H:i'),
-                'rights'=> $rightsForModel,
-            ],
-        ]);
-
-        Session::put('scout.rightsForModels', $sessionRecord);
+        return $model->getRightsForMandateTypes($mandateTypeIds, $isOwn, $is2fa, $ignoreCache);
     }
 
     public function isOwnModel($scout){
@@ -735,6 +714,20 @@ class Scout extends OrganizationBase
     public static function getOrganizationTypeModelNameUserFriendly()
     {
         return Lang::get('csatar.csatar::lang.plugin.admin.scout.scout');
+    }
+
+    public function getScoutOptions($scopes = null){
+        if (!empty($scopes['association']->value)) {
+            return self::associations(array_keys($scopes['association']->value))
+                ->select(Db::raw("concat(family_name, ' - ', given_name) as name, id"))
+                 ->lists('name', 'id')
+                ;
+        }
+        else {
+            return self::all()
+                ->select(Db::raw("concat(family_name, ' - ', given_name) as name, id"))
+                ->lists('name', 'id');
+        }
     }
 
     public function getStaticMessages(): array