+++ csatar-plugins/csatar/updates/SeederData.php
@@ -22,6 +22,7 @@ use Csatar\Csatar\Models\Training;
 use Csatar\Forms\Models\Form;
 use Seeder;
 use Db;
+use RainLab\User\Models\UserGroup;
 
 class SeederData extends Seeder
 {
@@ -198,6 +199,10 @@ class SeederData extends Seeder
             [
                 'title' => 'Képzés',
                 'model' => 'Csatar\Csatar\Models\Training',
+            ],
+            [
+                'title' => 'Baleseti log',
+                'model' => 'Csatar\Csatar\Models\AccidentLogRecord',
             ]
         ],
         'ageGroups' => [
@@ -405,6 +410,10 @@ class SeederData extends Seeder
         ],
         'sitesearchSettings' => [
             'enabledOnOrgCMSpages'
+        ],
+        'userGroups' => [
+            ['Accident log data entry', 'dataEntry'],
+            ['Accident log admin', 'admin']
         ],
     ];
 
@@ -636,6 +645,14 @@ class SeederData extends Seeder
                 $newAgeGroup->note = $ageGroup['note'];
                 $newAgeGroup->save();
             }
+        }
+
+        // userGroups
+        foreach($this::DATA['userGroups'] as $name) {
+            $allergy = UserGroup::firstOrCreate([
+                'name' => $name[0],
+                'code' => str_slug($name[0]),
+            ]);
         }
 
         // contact page data