+++ csatar-plugins/forms/traits/ManagesUploads.php
@@ -1,12 +1,12 @@
 <?php namespace Csatar\Forms\Traits;
 
+use File;
 use Input;
 use Lang;
+use October\Rain\Database\Models\DeferredBinding;
 use Request;
 use Response;
-use File;
 use Validator;
-use Session;
 
 // Returns a file size limit in bytes based on the PHP upload_max_filesize
 // and post_max_size
@@ -92,7 +92,24 @@ trait ManagesUploads {
 
     private function validateUpload($model_field) {
 
-        if (isset($this->record->rules[$model_field])) {
+        if (isset($this->record->rules[$model_field . '.*'])) {
+            $validationRules = $this->record->rules[$model_field . '.*'];
+            if(isset($this->record->rules[$model_field])) {
+                $arrayRule = $this->record->rules[$model_field];
+                // check if there is validation rule for $attachMany attachments and search for rule regarding max number of files
+                if ((preg_match('/max:(.*?)\|/', $arrayRule, $match) == 1
+                        || preg_match('/max:(.*?)$/', $arrayRule, $match))
+                    && (count($this->record->{$model_field}) >= $match[1]
+                        || $this->getNumberOfDeferredBindings($model_field, $this->sessionKey ) >= $match[1]
+                    )) {
+                        throw new \ValidationException([
+                            'attachments' => Lang::get('csatar.forms::lang.widgets.frontendFileUploadValidation.maxNumberOfAttachements',
+                                                                ['value' => $match[1]])
+                        ]);
+                    }
+            }
+        }
+        elseif (isset($this->record->rules[$model_field])) {
             $validationRules = $this->record->rules[$model_field];
         } else {
             $validationRules = ['max:' . (string) file_upload_max_size()];
@@ -158,8 +175,7 @@ trait ManagesUploads {
             $file->is_public = true;
             $file->save();
             if($isNew){
-                $sessionKey = Session::get('key');
-                $this->record->{$model_field}()->add($file, $sessionKey);
+                $this->record->{$model_field}()->add($file, $this->sessionKey);
             } else {
                 $this->record->{$model_field}()->add($file);
             }
@@ -300,6 +316,12 @@ trait ManagesUploads {
         }
 
         return $cssDimensions;
+    }
+
+    public function getNumberOfDeferredBindings($relationName, $sessionKey) {
+        return DeferredBinding::where('master_field', $relationName)
+                       ->where('session_key', $sessionKey)
+                       ->count();
     }
 
 }